Move into a packer_test pkg for black-box testing
