+++ packer_test/base_test.go
@@ -1,4 +1,4 @@
-package test
+package packer_test
 
 import (
 	"fmt"
