+++ packer_test/loading_test.go
@@ -1,4 +1,4 @@
-package test
+package packer_test
 
 import (
 	"crypto/sha256"
