+++ packer_test/mini_plugin/LICENSE
