+++ packer_test/mini_plugin/builder/dynamic/artifact.go
