+++ packer_test/mini_plugin/builder/dynamic/test-fixtures/template.pkr.hcl
