+++ packer_test/mini_plugin/datasource/dynamic/test-fixtures/template.pkr.hcl
