+++ packer_test/mini_plugin/example/README.md
