+++ packer_test/mini_plugin/example/build.pkr.hcl
