+++ packer_test/mini_plugin/example/data.pkr.hcl
