+++ packer_test/mini_plugin/example/variables.pkr.hcl
