+++ packer_test/mini_plugin/main.go
