+++ packer_test/mini_plugin/post-processor/dynamic/test-fixtures/template.pkr.hcl
