+++ packer_test/mini_plugin/provisioner/dynamic/test-fixtures/template.pkr.hcl
