+++ packer_test/templates/pin_1.0.9.pkr.hcl
