+++ packer_test/templates/simple.json
