+++ packer_test/templates/simple.pkr.hcl
